First version of actuator disk model implemented.
This required modifying the no-actuator-disk BEM method for including
a proper referential magnitude and a conditional for not solving the
inlet velocity triangle in case the input parameter already consists of
a partially solved turbine stage (or a second turbine stage for which the
inlet velocity triangle is equal to the outlet velocity triangle of the
previous stage).
It is to confirm whether the actuator disk model is well implemented.

BEM solvers and figure plotters put in different cells.

Deprecated code retrieved and dervied to the beta version of the notebook.